+++ sdk/formrecognizer/azure-ai-formrecognizer/src/main/java/com/azure/ai/formrecognizer/models/CustomFormSubmodel.java
@@ -5,6 +5,7 @@ package com.azure.ai.formrecognizer.models;
 
 import com.azure.core.annotation.Immutable;
 
+import java.util.Collections;
 import java.util.Map;
 
 /**
@@ -40,7 +41,7 @@ public final class CustomFormSubmodel {
     public CustomFormSubmodel(final Float accuracy, final Map<String, CustomFormModelField> fields,
         final String formType) {
         this.accuracy = accuracy;
-        this.fields = fields;
+        this.fields = fields == null ? null : Collections.unmodifiableMap(fields);
         this.formType = formType;
     }
 
@@ -67,7 +68,7 @@ public final class CustomFormSubmodel {
      * For models trained with labels, this is the training-time label of the field. For models trained with forms
      * only, a unique name is generated for each field.
      *
-     * @return the {@code fields} value
+     * @return the unmodifiable map of recognized fields.
      */
     public Map<String, CustomFormModelField> getFields() {
         return this.fields;