+++ sdk/resources/mgmt/src/main/java/com/azure/resourcemanager/resources/models/ResourceGroups.java
@@ -8,11 +8,11 @@ import com.azure.resourcemanager.resources.ResourceManager;
 import com.azure.resourcemanager.resources.fluentcore.arm.collection.SupportsGettingByName;
 import com.azure.resourcemanager.resources.fluentcore.arm.models.HasManager;
 import com.azure.resourcemanager.resources.fluentcore.collection.SupportsBatchCreation;
-import com.azure.resourcemanager.resources.fluentcore.collection.SupportsBeginDeletingByName;
 import com.azure.resourcemanager.resources.fluentcore.collection.SupportsCreating;
 import com.azure.resourcemanager.resources.fluentcore.collection.SupportsDeletingByName;
 import com.azure.resourcemanager.resources.fluentcore.collection.SupportsListing;
 import com.azure.resourcemanager.resources.fluentcore.collection.SupportsListingByTag;
+import com.azure.resourcemanager.resources.fluentcore.model.Accepted;
 
 /**
  * Entry point to resource group management API.
@@ -24,7 +24,7 @@ public interface ResourceGroups extends
         SupportsGettingByName<ResourceGroup>,
         SupportsCreating<ResourceGroup.DefinitionStages.Blank>,
         SupportsDeletingByName,
-        SupportsBeginDeletingByName,
+        //SupportsBeginDeletingByName,
         SupportsBatchCreation<ResourceGroup>,
         HasManager<ResourceManager> {
 
@@ -35,4 +35,12 @@ public interface ResourceGroups extends
      * @return true of exists, otherwise false
      */
     boolean contain(String name);
+
+    /**
+     * Begins deleting a resource group from Azure, identifying it by its name.
+     *
+     * @param name the resource group name
+     * @return the accepted deleting operation
+     */
+    Accepted<Void> beginDeleteByName(String name);
 }