+++ sdk/search/azure-search-documents/src/main/java/com/azure/search/documents/implementation/converters/SuggestResultConverter.java
@@ -7,8 +7,6 @@ import com.azure.search.documents.SearchDocument;
 import com.azure.search.documents.implementation.util.PrivateFieldAccessHelper;
 import com.azure.search.documents.models.SuggestResult;
 
-import java.util.Map;
-
 /**
  * A converter between {@link com.azure.search.documents.implementation.models.SuggestResult} and {@link SuggestResult}.
  */
@@ -40,7 +38,7 @@ public final class SuggestResultConverter {
         com.azure.search.documents.implementation.models.SuggestResult suggestResult =
             new com.azure.search.documents.implementation.models.SuggestResult();
 
-        Map<String, Object> additionalProperties = obj.getDocument();
+        SearchDocument additionalProperties = obj.getDocument(SearchDocument.class);
         PrivateFieldAccessHelper.set(suggestResult, "additionalProperties", additionalProperties);
 
         String text = obj.getText();