+++ sdk/storage/azure-storage-file-datalake/src/test/java/com/azure/storage/file/datalake/FileAPITest.groovy
@@ -2711,5 +2711,4 @@ class FileAPITest extends APISpec {
         cleanup:
         file.delete()
     }
-
 }