+++ sdk/core/azure-core/CHANGELOG.md
@@ -1,7 +1,7 @@
 # Release History
 
 ## 1.6.0-beta.1 (Unreleased)
-
+- Added `TokenRefreshOptions()` to `TokenCredential`, with a default token refresh offset of 2 minutes, and a default token refresh retry timeout of 30 seconds.
 
 ## 1.5.1 (2020-06-08)